requirements.txt
python-dotenv
beautifulsoup4
requests
json
gradio==3.32.0
setuptools
Pillow

